requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
finnhub-python==2.4.21
plotly==5.23.0
yfinance==0.2.40
feedparser==6.0.11
